+++ src/main/java/homeTry/team/dto/response/TeamResponse.java
@@ -1,4 +1,4 @@
-package homeTry.team.dto;
+package homeTry.team.dto.response;
 
 import homeTry.tag.dto.TagDTO;
 import homeTry.team.model.entity.Team;
